Move FormPanel class to common directory.
